backend api/account/views.py
from rest_framework_simplejwt.tokens import RefreshToken
from rest_framework import generics, status
from rest_framework.permissions import AllowAny
from django.contrib.auth import authenticate
from rest_framework.response import Response
from .serializers import UserLoginSerializer

class JWTLoginView(generics.GenericAPIView):
    permission_classes = [AllowAny] 
    serializer_class = UserLoginSerializer

    def post(self, request, *args, **kwargs):
        email = request.data.get('email')
        password = request.data.get('password')
        terms_conditions = request.data.get('terms_conditions', False) 

        if not terms_conditions:
            return Response({'error': 'You must agree to the terms and conditions.'}, status=status.HTTP_400_BAD_REQUEST)

        # Authenticate the user using email and password
        user = authenticate(email=email, password=password)

        if user is not None:
            refresh = RefreshToken.for_user(user)
            return Response({
                'refresh': str(refresh),
                'access': str(refresh.access_token),
            }, status=status.HTTP_200_OK)
        else:
            return Response({'error': 'Invalid Credentials'}, status=status.HTTP_401_UNAUTHORIZED)
