backend api/pyproject.toml
[project]
name = "installment-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cloudinary>=1.41.0",
    "django-cors-headers>=4.6.0",
    "django>=5.1.3",
    "django-filter>=24.3",
    "djangorestframework>=3.15.2",
    "djangorestframework-simplejwt>=5.3.1",
    "pillow>=11.0.0",
    "python-decouple>=3.8",
]
